+++ R/srStarts.R
@@ -177,7 +177,8 @@ iSRStartsI <- function(S,R) c(a=stats::coef(stats::lm(R~0+S))[[1]])
 #=============================================================
 # Static plot of starting values
 #=============================================================
-iSRStartsPlot <- function(S,R,type,param,sv,col.mdl,lwd.mdl,lty.mdl,cex.main,col.main) {
+iSRStartsPlot <- function(S,R,type,param,sv,col.mdl,lwd.mdl,lty.mdl,
+                          cex.main,col.main) { # nocov start
   ## attempting to get by bindings warning in RCMD CHECK
   x <- NULL
   ## Make a title
@@ -192,4 +193,4 @@ iSRStartsPlot <- function(S,R,type,param,sv,col.mdl,lwd.mdl,lty.mdl,cex.main,col
   graphics::curve(mdl(x,unlist(sv)),from=0,to=max.S,col=col.mdl,lwd=lwd.mdl,lty=lty.mdl,add=TRUE)
   legend("topleft",paste(names(sv),formatC(unlist(sv),format="f",digits=3),sep="="),
          bty="n",cex=cex.main)
-}
+} # nocov end