+++ R/stockRecruitment.R
@@ -217,10 +217,10 @@ srFuns <- function(type=c("BevertonHolt","Ricker","Shepherd","SailaLorda","indep
 
 #' @rdname stockRecruitment
 #' @export
-srModels <- function(...) {
+srModels <- function(...) { # nocov start
   op <- graphics::par(mar=c(0,0,2,0))
-  graphics::plot(1,type="n",ylim=c(0,6),xlim=c(0,1),xaxt="n",yaxt="n",xlab="",ylab="",bty="n",
-       main="FSA Stock-Recruit Model Parametrizations",...)
+  graphics::plot(1,type="n",ylim=c(0,6),xlim=c(0,1),xaxt="n",yaxt="n",
+                 xlab="",ylab="",bty="n",main="FSA Stock-Recruit Model Parametrizations",...)
   iSRModels("BH1",0,5.5)
   iSRModels("BH2",0,4.25)
   iSRModels("BH3",0,3)
@@ -232,10 +232,10 @@ srModels <- function(...) {
   iSRModels("Shepherd",0.55,1.75)
   iSRModels("SailaLorda",0.55,0.5)
   graphics::par(op)
-}
+} # nocov end
 
 ## Internal function for plotting the different models.  Send positions in xpos and ypos.
-iSRModels <- function(which,xpos,ypos) {
+iSRModels <- function(which,xpos,ypos) { # nocov start
   switch(which,
          BH1={graphics::text(xpos,ypos,expression(plain("BevertonHolt #1: ")~~~R==frac(aS,1+bS)),pos=4)},
          BH2={graphics::text(xpos,ypos,expression(plain("BevertonHolt #2: ")~~~R==frac(aS,1+a*~frac(S,R[p]))),pos=4)},
@@ -247,4 +247,4 @@ iSRModels <- function(which,xpos,ypos) {
          Shepherd={graphics::text(xpos,ypos,expression(plain("Shepherd: ")~~~R==frac(aS,1+(bS)^{c})),pos=4)},
          SailaLorda={graphics::text(xpos,ypos,expression(plain("Saila-Lorda: ")~~~R==aS^{c}*e^{-bS}),pos=4)}
   )
-}
+} # nocov end